+++ scripts/start_smartbike.sh
@@ -1,0 +1,3 @@
+source ~/.env
+
+python3 ~/iot/Drivers/smartbike_driver/smartbike.py &
